+++ src/commands/browser/stagehand/initOverride.ts
@@ -285,16 +285,16 @@ export const patchStagehand = once(async () => {
   return { StagehandPage, StagehandContext };
 });
 
+export type StagehandInitOverride = {
+  recordVideo: RecordVideoOptions | undefined;
+  viewport?: SharedBrowserCommandOptions['viewport'];
+  connectTo?: SharedBrowserCommandOptions['connectTo'];
+};
+
 export function overrideStagehandInit() {
   //@ts-ignore
   Stagehand.prototype.init = async function myinit(
-    initOptions:
-      | undefined
-      | {
-          recordVideo: RecordVideoOptions;
-          viewport?: SharedBrowserCommandOptions['viewport'];
-          connectTo?: SharedBrowserCommandOptions['connectTo'];
-        }
+    initOptions: undefined | StagehandInitOverride
   ): Promise<{ debugUrl: string; sessionUrl: string; sessionId: string }> {
     try {
       const { StagehandPage, StagehandContext } = await patchStagehand();